Cleans up a few thigns
The timing for the enemy attack needed to be reigned in a little. The player's raycast was pointed in the wrong direction. On level 2, there was a hidden staticbody that needed to be removed. I cleaned up a few other things.

Hopefully that helps. Let me know what else I can do!